requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
sqlparse==0.5.0
tablib==3.5.0
